Updated GPU stuff

+++ trunk/HW/termProject/debug.h
@@ -2,10 +2,11 @@
 #define DEBUG_H
 
 #include <iostream>
+#include <stdint.h>
 #include "mass.h" 
 
 // For debugging
-void printTime(int seconds);
+void printTime(uint64_t seconds);
 void printF(Mass m);
 void printVel(Mass m);
 void printPos(Mass m);
