+++ trunk/HW/termProject/output.h
@@ -10,9 +10,9 @@ extern FILE * ext_vidFile;
 
 void initOutput();
 void initOutput(char * name);
-void outputObjectCount(uint32_t objectCount);
+void outputObjectCount(uint64_t objectCount);
 void outputObjectMass(double mass);
-void outputSimSteps(uint32_t simSteps);
+void outputSimSteps(uint64_t simSteps);
 void outputFrames(double x, double y); 
 void outputClose(); 
 void outputResults(vector<Mass> massVector);
